+++ src/plugins.ts
@@ -4,8 +4,8 @@ import type Eslint from 'eslint';
 import ourPackageJson from '../package.json' with {type: 'json'};
 import type {UnConfigContext} from './configs';
 import type {EslintPlugin} from './eslint';
-import type {Promisable} from './types';
-import {type MaybeArray, arraify, interopDefault, objectKeysUnsafe, omit} from './utils';
+import type {ObjectValues, Promisable} from './types';
+import {type MaybeArray, arraify, interopDefault, isIn, objectKeysUnsafe, omit} from './utils';
 
 export const OPTIONAL_PLUGINS_PACKAGE_NAMES = omit(ourPackageJson.peerDependencies, ['eslint']);
 
@@ -15,37 +15,47 @@ interface LoadPluginOptions {
   throwIfNotFound?: boolean;
 }
 
-export function genPluginLoader<T, N extends string>(
+type PluginLoader<
+  T,
+  Prefix extends string,
+  N extends string,
+  PackageNullable extends boolean = true,
+> = Record<
+  Prefix,
+  (
+    context: UnConfigContext,
+    options?: LoadPluginOptions,
+  ) => Promise<{
+    packageName: N;
+    module: T | (PackageNullable extends true ? null : never);
+  }>
+>;
+
+export function genPluginLoader<T, Prefix extends string, N extends string>(
+  prefix: Prefix,
   packageName: N,
   module: () => Promisable<T | {default: T}>,
   ignoreErrors?: undefined,
-): (
-  context: UnConfigContext,
-  options?: LoadPluginOptions,
-) => Promise<{
-  packageName: N;
-  contents: T | (N extends keyof typeof OPTIONAL_PLUGINS_PACKAGE_NAMES ? null : never);
-}>;
-export function genPluginLoader<T, N extends string>(
+): PluginLoader<T, Prefix, N, N extends keyof typeof OPTIONAL_PLUGINS_PACKAGE_NAMES ? true : false>;
+export function genPluginLoader<T, Prefix extends string, N extends string>(
+  prefix: Prefix,
   packageName: N,
   module: () => Promisable<T | {default: T}>,
   ignoreErrors: MaybeArray<string>,
-): (
-  context: UnConfigContext,
-  options?: LoadPluginOptions,
-) => Promise<{packageName: N; contents: T | null}>;
-export function genPluginLoader<T, N extends string>(
+): PluginLoader<T, Prefix, N>;
+export function genPluginLoader<T, Prefix extends string, N extends string>(
+  prefix: Prefix,
   packageName: N,
   module: () => Promisable<T | {default: T}>,
   ignoreErrors?: MaybeArray<string>,
-): (
-  context: UnConfigContext,
-  options?: LoadPluginOptions,
-) => Promise<{packageName: N; contents: T | null}> {
-  return async (context, options) => {
+): PluginLoader<T, Prefix, N> {
+  const loader: PluginLoader<T, Prefix, N>[Prefix] = async (context, options) => {
     const isPluginOptionalPeerDependency = packageName in OPTIONAL_PLUGINS_PACKAGE_NAMES;
     try {
-      return {contents: await interopDefault(module()), packageName};
+      const {pluginsOverrides} = context.rootOptions;
+      const providedPlugin =
+        pluginsOverrides && isIn(prefix, pluginsOverrides) ? (pluginsOverrides[prefix] as T) : null;
+      return {module: providedPlugin || (await interopDefault(module())), packageName};
     } catch (error: unknown) {
       const ignoredErrors: string[] = [
         ...arraify(ignoreErrors),
@@ -59,99 +69,123 @@ export function genPluginLoader<T, N extends string>(
         ignoredErrors.includes(error.code) &&
         !options?.throwIfNotFound
       ) {
-        return {contents: null, packageName};
+        return {module: null, packageName};
       }
       throw error;
     }
   };
+  return {
+    [prefix]: loader,
+  } as PluginLoader<T, Prefix, N>;
 }
 
 export const pluginsLoaders = {
-  '@cspell': genPluginLoader('@cspell/eslint-plugin', () => import('@cspell/eslint-plugin')),
-  '@eslint-community/eslint-comments': genPluginLoader(
+  ...genPluginLoader('@cspell', '@cspell/eslint-plugin', () => import('@cspell/eslint-plugin')),
+  ...genPluginLoader(
+    '@eslint-community/eslint-comments',
     '@eslint-community/eslint-plugin-eslint-comments',
     () => import('@eslint-community/eslint-plugin-eslint-comments'),
   ),
-  '@eslint-react': genPluginLoader('@eslint-react/eslint-plugin', () =>
+  ...genPluginLoader('@eslint-react', '@eslint-react/eslint-plugin', () =>
     interopDefault(import('@eslint-react/eslint-plugin')).then(
       (m) => m.configs.all.plugins['@eslint-react'] as unknown as EslintPlugin,
     ),
   ),
-  '@eslint-react/debug': genPluginLoader('@eslint-react/eslint-plugin', () =>
+  ...genPluginLoader('@eslint-react/debug', '@eslint-react/eslint-plugin', () =>
     interopDefault(import('@eslint-react/eslint-plugin')).then(
       (m) => m.configs.all.plugins['@eslint-react/debug'] as unknown as EslintPlugin,
     ),
   ),
-  '@eslint-react/dom': genPluginLoader('@eslint-react/eslint-plugin', () =>
+  ...genPluginLoader('@eslint-react/dom', '@eslint-react/eslint-plugin', () =>
     interopDefault(import('@eslint-react/eslint-plugin')).then(
       (m) => m.configs.all.plugins['@eslint-react/dom'] as unknown as EslintPlugin,
     ),
   ),
-  '@eslint-react/hooks-extra': genPluginLoader('@eslint-react/eslint-plugin', () =>
+  ...genPluginLoader('@eslint-react/hooks-extra', '@eslint-react/eslint-plugin', () =>
     interopDefault(import('@eslint-react/eslint-plugin')).then(
       (m) => m.configs.all.plugins['@eslint-react/hooks-extra'] as unknown as EslintPlugin,
     ),
   ),
-  '@eslint-react/naming-convention': genPluginLoader('@eslint-react/eslint-plugin', () =>
+  ...genPluginLoader('@eslint-react/naming-convention', '@eslint-react/eslint-plugin', () =>
     interopDefault(import('@eslint-react/eslint-plugin')).then(
       (m) => m.configs.all.plugins['@eslint-react/naming-convention'] as unknown as EslintPlugin,
     ),
   ),
-  '@eslint-react/web-api': genPluginLoader('@eslint-react/eslint-plugin', () =>
+  ...genPluginLoader('@eslint-react/web-api', '@eslint-react/eslint-plugin', () =>
     interopDefault(import('@eslint-react/eslint-plugin')).then(
       (m) => m.configs.all.plugins['@eslint-react/web-api'] as unknown as EslintPlugin,
     ),
   ),
-  '@html-eslint': genPluginLoader(
+  ...genPluginLoader(
+    '@html-eslint',
     '@html-eslint/eslint-plugin',
     () => import('@html-eslint/eslint-plugin'),
   ),
-  '@next/next': genPluginLoader(
+  ...genPluginLoader(
+    '@next/next',
     '@next/eslint-plugin-next',
     () => import('@next/eslint-plugin-next'),
   ),
   // We can't `import()` `@stylistic/eslint-plugin` because it's `require()`d by eslint-plugin-vue: https://github.com/vuejs/eslint-plugin-vue/blob/1b634549a9e91231e5ea79313763c69f93e678c1/lib/utils/index.js#L113 and `import()`ing after `require()`ing causes `ERR_INTERNAL_ASSERTION` error, see https://github.com/nodejs/node/issues/54577
-  '@stylistic': genPluginLoader('@stylistic', () => Promise.resolve(stylistic)),
-  // @ts-expect-error types mismatch
-  '@tanstack/query': genPluginLoader(
+  ...genPluginLoader('@stylistic', '@stylistic', () => Promise.resolve(stylistic)),
+  ...genPluginLoader(
+    '@tanstack/query',
     '@tanstack/eslint-plugin-query',
-    () => import('@tanstack/eslint-plugin-query'),
+    () =>
+      import(
+        '@tanstack/eslint-plugin-query'
+        // @ts-expect-error types mismatch
+      ) satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
   ),
-  astro: genPluginLoader('eslint-plugin-astro', () => import('eslint-plugin-astro')),
-  ava: genPluginLoader('eslint-plugin-ava', () => import('eslint-plugin-ava')),
-  'better-tailwindcss': genPluginLoader(
+  ...genPluginLoader('astro', 'eslint-plugin-astro', () => import('eslint-plugin-astro')),
+  ...genPluginLoader('ava', 'eslint-plugin-ava', () => import('eslint-plugin-ava')),
+  ...genPluginLoader(
+    'better-tailwindcss',
     'eslint-plugin-better-tailwindcss',
     () => import('eslint-plugin-better-tailwindcss'),
   ),
-  'case-police': genPluginLoader(
+  ...genPluginLoader(
+    'case-police',
     'eslint-plugin-case-police',
     () => import('eslint-plugin-case-police') as Promise<EslintPlugin>,
   ),
-  css: genPluginLoader('@eslint/css', () => import('@eslint/css')),
-  // @ts-expect-error types mismatch
-  'css-in-js': genPluginLoader('eslint-plugin-css', () => import('eslint-plugin-css')),
-  cypress: genPluginLoader('eslint-plugin-cypress', () => import('eslint-plugin-cypress')),
-  'de-morgan': genPluginLoader(
+  ...genPluginLoader('css', '@eslint/css', () => import('@eslint/css')),
+  ...genPluginLoader(
+    'css-in-js',
+    'eslint-plugin-css',
+    // @ts-expect-error types mismatch
+    () => import('eslint-plugin-css') satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
+  ),
+  ...genPluginLoader('cypress', 'eslint-plugin-cypress', () => import('eslint-plugin-cypress')),
+  ...genPluginLoader(
+    'de-morgan',
     'eslint-plugin-de-morgan',
     () => import('eslint-plugin-de-morgan') as Promise<EslintPlugin>,
   ),
-  depend: genPluginLoader('eslint-plugin-depend', () => import('eslint-plugin-depend')),
-  ember: genPluginLoader('eslint-plugin-ember', () => import('eslint-plugin-ember')),
-  // @ts-expect-error types mismatch
-  'erasable-syntax-only': genPluginLoader(
+  ...genPluginLoader('depend', 'eslint-plugin-depend', () => import('eslint-plugin-depend')),
+  ...genPluginLoader('ember', 'eslint-plugin-ember', () => import('eslint-plugin-ember')),
+  ...genPluginLoader(
+    'erasable-syntax-only',
     'eslint-plugin-erasable-syntax-only',
-    () => import('eslint-plugin-erasable-syntax-only'),
+    () =>
+      import(
+        'eslint-plugin-erasable-syntax-only'
+        // @ts-expect-error types mismatch
+      ) satisfies Promise<EslintPlugin> as unknown as Promise<EslintPlugin>,
   ),
-  es: genPluginLoader('eslint-plugin-es-x', () => import('eslint-plugin-es-x')),
-  'eslint-plugin': genPluginLoader(
+  ...genPluginLoader('es', 'eslint-plugin-es-x', () => import('eslint-plugin-es-x')),
+  ...genPluginLoader(
+    'eslint-plugin',
     'eslint-plugin-eslint-plugin',
     () => import('eslint-plugin-eslint-plugin'),
   ),
-  'file-progress': genPluginLoader(
+  ...genPluginLoader(
+    'file-progress',
     'eslint-plugin-file-progress',
     () => import('eslint-plugin-file-progress'),
   ),
-  graphql: genPluginLoader(
+  ...genPluginLoader(
+    'graphql',
     '@graphql-eslint/eslint-plugin',
     () =>
       import('@graphql-eslint/eslint-plugin') as Promise<{
@@ -161,133 +195,181 @@ export const pluginsLoaders = {
         };
       }>,
   ),
-  html: genPluginLoader('eslint-plugin-html', () => import('eslint-plugin-html')),
-  import: genPluginLoader(
+  ...genPluginLoader('html', 'eslint-plugin-html', () => import('eslint-plugin-html')),
+  ...genPluginLoader(
+    'import',
     'eslint-plugin-import-x',
     () => import('eslint-plugin-import-x') as unknown as Promise<EslintPlugin>,
   ),
-  jest: genPluginLoader('eslint-plugin-jest', () => import('eslint-plugin-jest')),
-  'jest-extended': genPluginLoader(
+  ...genPluginLoader('jest', 'eslint-plugin-jest', () => import('eslint-plugin-jest')),
+  ...genPluginLoader(
+    'jest-extended',
     'eslint-plugin-jest-extended',
     () => import('eslint-plugin-jest-extended'),
   ),
-  jsdoc: genPluginLoader('eslint-plugin-jsdoc', () => import('eslint-plugin-jsdoc')),
-  // @ts-expect-error types mismatch
-  'json-schema-validator': genPluginLoader(
+  ...genPluginLoader('jsdoc', 'eslint-plugin-jsdoc', () => import('eslint-plugin-jsdoc')),
+  ...genPluginLoader(
+    'json-schema-validator',
     'eslint-plugin-json-schema-validator',
-    () => import('eslint-plugin-json-schema-validator'),
+    () =>
+      import(
+        'eslint-plugin-json-schema-validator'
+        // @ts-expect-error types mismatch
+      ) satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
   ),
-  // @ts-expect-error types mismatch
-  jsonc: genPluginLoader('eslint-plugin-jsonc', () => import('eslint-plugin-jsonc')),
-  'jsx-a11y': genPluginLoader('eslint-plugin-jsx-a11y', () => import('eslint-plugin-jsx-a11y')),
-  markdown: genPluginLoader('@eslint/markdown', () => import('@eslint/markdown')),
-  // @ts-expect-error types mismatch
-  math: genPluginLoader<EslintPlugin>('eslint-plugin-math', () => import('eslint-plugin-math')),
-  mdx: genPluginLoader('eslint-plugin-mdx', () => import('eslint-plugin-mdx')),
-  'no-type-assertion': genPluginLoader(
+  ...genPluginLoader(
+    'jsonc',
+    'eslint-plugin-jsonc',
+    // @ts-expect-error types mismatch
+    () => import('eslint-plugin-jsonc') satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
+  ),
+  ...genPluginLoader('jsx-a11y', 'eslint-plugin-jsx-a11y', () => import('eslint-plugin-jsx-a11y')),
+  ...genPluginLoader('markdown', '@eslint/markdown', () => import('@eslint/markdown')),
+  ...genPluginLoader(
+    'math',
+    'eslint-plugin-math',
+    // @ts-expect-error types mismatch
+    () => import('eslint-plugin-math') satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
+  ),
+  ...genPluginLoader('mdx', 'eslint-plugin-mdx', () => import('eslint-plugin-mdx')),
+  ...genPluginLoader(
+    'no-type-assertion',
     'eslint-plugin-no-type-assertion',
     () => import('eslint-plugin-no-type-assertion'),
   ),
-  'no-unsanitized': genPluginLoader(
+  ...genPluginLoader(
+    'no-unsanitized',
     'eslint-plugin-no-unsanitized',
     () => import('eslint-plugin-no-unsanitized'),
   ),
-  node: genPluginLoader('eslint-plugin-n', () => import('eslint-plugin-n')),
-  // @ts-expect-error types mismatch
-  'node-dependencies': genPluginLoader(
+  ...genPluginLoader('node', 'eslint-plugin-n', () => import('eslint-plugin-n')),
+  ...genPluginLoader(
+    'node-dependencies',
     'eslint-plugin-node-dependencies',
-    () => import('eslint-plugin-node-dependencies'),
+    () =>
+      import(
+        'eslint-plugin-node-dependencies'
+        // @ts-expect-error types mismatch
+      ) satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
   ),
-  // @ts-expect-error types mismatch
-  'package-json': genPluginLoader(
+  ...genPluginLoader(
+    'package-json',
     'eslint-plugin-package-json',
-    () => import('eslint-plugin-package-json'),
+    () =>
+      // @ts-expect-error types mismatch
+      import('eslint-plugin-package-json') satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
   ),
-  perfectionist: genPluginLoader(
+  ...genPluginLoader(
+    'perfectionist',
     'eslint-plugin-perfectionist',
     () => import('eslint-plugin-perfectionist') as Promise<EslintPlugin>,
   ),
-  // @ts-expect-error types mismatch
-  pinia: genPluginLoader('eslint-plugin-pinia', () => import('eslint-plugin-pinia')),
-  playwright: genPluginLoader('eslint-plugin-playwright', () => import('eslint-plugin-playwright')),
-  pnpm: genPluginLoader('eslint-plugin-pnpm', () => import('eslint-plugin-pnpm')),
-  // @ts-expect-error types mismatch
-  'prefer-arrow-functions': genPluginLoader(
-    'eslint-plugin-prefer-arrow-functions',
-    () => import('eslint-plugin-prefer-arrow-functions'),
+  ...genPluginLoader(
+    'pinia',
+    'eslint-plugin-pinia',
+    // @ts-expect-error types mismatch
+    () => import('eslint-plugin-pinia') satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
   ),
-  prettier: genPluginLoader('eslint-plugin-prettier', () => import('eslint-plugin-prettier')),
-  promise: genPluginLoader('eslint-plugin-promise', () => import('eslint-plugin-promise')),
-  qwik: genPluginLoader('eslint-plugin-qwik', () =>
+  ...genPluginLoader(
+    'playwright',
+    'eslint-plugin-playwright',
+    () => import('eslint-plugin-playwright'),
+  ),
+  ...genPluginLoader('pnpm', 'eslint-plugin-pnpm', () => import('eslint-plugin-pnpm')),
+  ...genPluginLoader(
+    'prefer-arrow-functions',
+    'eslint-plugin-prefer-arrow-functions',
+    () =>
+      import(
+        'eslint-plugin-prefer-arrow-functions'
+        // @ts-expect-error types mismatch
+      ) satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
+  ),
+  ...genPluginLoader('prettier', 'eslint-plugin-prettier', () => import('eslint-plugin-prettier')),
+  ...genPluginLoader('promise', 'eslint-plugin-promise', () => import('eslint-plugin-promise')),
+  ...genPluginLoader('qwik', 'eslint-plugin-qwik', () =>
     interopDefault(import('eslint-plugin-qwik')).then((m) => fixupPluginRules(m)),
   ),
-  react: genPluginLoader('eslint-plugin-react', () => import('eslint-plugin-react')),
-  'react-compiler': genPluginLoader(
+  ...genPluginLoader('react', 'eslint-plugin-react', () => import('eslint-plugin-react')),
+  ...genPluginLoader(
+    'react-compiler',
     'eslint-plugin-react-compiler',
     () => import('eslint-plugin-react-compiler'),
   ),
-  'react-hooks': genPluginLoader(
+  ...genPluginLoader(
+    'react-hooks',
     'eslint-plugin-react-hooks',
     () => import('eslint-plugin-react-hooks'),
   ),
-  'react-refresh': genPluginLoader(
+  ...genPluginLoader(
+    'react-refresh',
     'eslint-plugin-react-refresh',
     () => import('eslint-plugin-react-refresh'),
   ),
-  regexp: genPluginLoader('eslint-plugin-regexp', () => import('eslint-plugin-regexp')),
-  security: genPluginLoader('eslint-plugin-security', () => import('eslint-plugin-security')),
-  solid: genPluginLoader(
+  ...genPluginLoader('regexp', 'eslint-plugin-regexp', () => import('eslint-plugin-regexp')),
+  ...genPluginLoader('security', 'eslint-plugin-security', () => import('eslint-plugin-security')),
+  ...genPluginLoader(
+    'solid',
     'eslint-plugin-solid',
     () => import('eslint-plugin-solid') as Promise<EslintPlugin>,
   ),
-  sonarjs: genPluginLoader('eslint-plugin-sonarjs', () => import('eslint-plugin-sonarjs')),
-  storybook: genPluginLoader(
+  ...genPluginLoader('sonarjs', 'eslint-plugin-sonarjs', () => import('eslint-plugin-sonarjs')),
+  ...genPluginLoader(
+    'storybook',
     'eslint-plugin-storybook',
     () => import('eslint-plugin-storybook') as Promise<EslintPlugin>,
   ),
-  svelte: genPluginLoader(
+  ...genPluginLoader(
+    'svelte',
     'eslint-plugin-svelte',
     () => import('eslint-plugin-svelte'),
     // Hard-depends on `svelte` package, uses it at least in `lib/utils/svelte-context.js`
     MODULE_NOT_FOUND_ERROR_CODES,
   ),
-  tailwindcss: genPluginLoader(
+  ...genPluginLoader(
+    'tailwindcss',
     'eslint-plugin-tailwindcss',
     () => import('eslint-plugin-tailwindcss'),
     // Tries to import `tailwindcss/resolveConfig` which doesn't exist anymore in v4
     ['ERR_PACKAGE_PATH_NOT_EXPORTED', ...MODULE_NOT_FOUND_ERROR_CODES],
   ),
-  'testing-library': genPluginLoader(
+  ...genPluginLoader(
+    'testing-library',
     'eslint-plugin-testing-library',
     () => import('eslint-plugin-testing-library'),
   ),
-  // @ts-expect-error types mismatch
-  toml: genPluginLoader('eslint-plugin-toml', () => import('eslint-plugin-toml')),
-  ts: genPluginLoader(
+  ...genPluginLoader(
+    'toml',
+    'eslint-plugin-toml',
+    // @ts-expect-error types mismatch
+    () => import('eslint-plugin-toml') satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
+  ),
+  ...genPluginLoader(
+    'ts',
     'typescript-eslint',
     () => import('typescript-eslint').then((m) => m.plugin) as Promise<EslintPlugin>,
   ),
-  turbo: genPluginLoader('eslint-plugin-turbo', () => import('eslint-plugin-turbo')),
-  unicorn: genPluginLoader('eslint-plugin-unicorn', () => import('eslint-plugin-unicorn')),
-  'unused-imports': genPluginLoader(
+  ...genPluginLoader('turbo', 'eslint-plugin-turbo', () => import('eslint-plugin-turbo')),
+  ...genPluginLoader('unicorn', 'eslint-plugin-unicorn', () => import('eslint-plugin-unicorn')),
+  ...genPluginLoader(
+    'unused-imports',
     'eslint-plugin-unused-imports',
     () => import('eslint-plugin-unused-imports'),
   ),
-  vitest: genPluginLoader('@vitest/eslint-plugin', () => import('@vitest/eslint-plugin')),
-  vue: genPluginLoader('eslint-plugin-vue', () => import('eslint-plugin-vue')),
-  'vuejs-accessibility': genPluginLoader(
+  ...genPluginLoader('vitest', '@vitest/eslint-plugin', () => import('@vitest/eslint-plugin')),
+  ...genPluginLoader('vue', 'eslint-plugin-vue', () => import('eslint-plugin-vue')),
+  ...genPluginLoader(
+    'vuejs-accessibility',
     'eslint-plugin-vuejs-accessibility',
     () => import('eslint-plugin-vuejs-accessibility'),
   ),
-  // @ts-expect-error types mismatch
-  yml: genPluginLoader('eslint-plugin-yml', () => import('eslint-plugin-yml')),
-} satisfies Record<
-  string,
-  (
-    context: UnConfigContext,
-    options?: LoadPluginOptions,
-  ) => Promise<{packageName: string; contents: EslintPlugin | null}>
->;
+  ...genPluginLoader(
+    'yml',
+    'eslint-plugin-yml',
+    // @ts-expect-error types mismatch
+    () => import('eslint-plugin-yml') satisfies Promise<EslintPlugin> as Promise<EslintPlugin>,
+  ),
+} satisfies Record<string, ObjectValues<PluginLoader<EslintPlugin, string, string>>>;
 
 type LoadablePluginPrefix = keyof typeof pluginsLoaders;
 export const LOADABLE_PLUGIN_PREFIXES_LIST = objectKeysUnsafe(pluginsLoaders);